Some update to UI

+++ QuanLyPMT/GUI/QuanLyBenhNhan.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.Globalization;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using BUS;
@@ -35,9 +37,15 @@ namespace GUI
         }
         private void deleteBtn_Click(object sender, EventArgs e)
         {
-            benhNhan_BUS.deleteBenhNhan(fromTable.MABENHNHAN);
-            MessageBox.Show("Xoá thông tin thành công");
-            showData();
+            if (MessageBox.Show
+               ("Bạn có chắc chắn muốn xóa thông tin này, dữ liệu có thể sẽ không được khôi phục",
+                "Thông báo",
+                MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == System.Windows.Forms.DialogResult.OK)
+            {
+                benhNhan_BUS.deleteBenhNhan(fromTable.MABENHNHAN);
+                MessageBox.Show("Xoá thông tin thành công","Thông báo",MessageBoxButtons.OK, MessageBoxIcon.Information);
+                showData();
+            }
         }
 
         private void danhSachBenhNhanDgv_RowEnter(object sender, DataGridViewCellEventArgs e)
@@ -53,6 +61,7 @@ namespace GUI
                 fromTable.HOTEN = danhSachBenhNhanDgv.Rows[index].Cells["HOTEN"].Value.ToString();
                 fromTable.DIACHI = danhSachBenhNhanDgv.Rows[index].Cells["DIACHI"].Value.ToString();
                 fromTable.SDT = danhSachBenhNhanDgv.Rows[index].Cells["SDT"].Value.ToString();
+                fromTable.GHICHU =  danhSachBenhNhanDgv.Rows[index].Cells["GHICHU"].Value.ToString();
             }
             setValueToForm();
         }
@@ -61,9 +70,13 @@ namespace GUI
             CMNDtextBOX1.Text = fromTable.CMND;
             CMNDtextBox2.Text = fromTable.CMND;
             nameTxtBox.Text = fromTable.HOTEN;
-            if(fromTable.GIOITINH=="Nữ")
+            if (fromTable.GIOITINH == "Nữ")
             {
                 femaleBtn.Checked = true;
+            }
+            else
+            {
+                femaleBtn.Checked = false;
             }
             birthDateTimePicker.Text = fromTable.NGAYSINH.ToString();
             addressTextbox.Text = fromTable.DIACHI;
@@ -72,9 +85,9 @@ namespace GUI
         }
         private void setValuefromForm()
         {
-           
+
             fromForm.CMND = CMNDtextBOX1.Text;
-            
+
             fromForm.HOTEN = nameTxtBox.Text;
             if (femaleBtn.Checked)
             {
@@ -91,6 +104,7 @@ namespace GUI
         }
         private void adddBtn_Click(object sender, EventArgs e)
         {
+            clearForm();
             isUsingAdd = true;
             isUsingUpdate = false;
             adddBtn.Visible = false;
@@ -98,6 +112,9 @@ namespace GUI
             updateBtn.Visible = false;
             saveBtn.Visible = true;
             cancelBtn.Visible = true;
+
+            unloadForm();
+
         }
 
         private void updateBtn_Click(object sender, EventArgs e)
@@ -109,63 +126,93 @@ namespace GUI
             updateBtn.Visible = false;
             saveBtn.Visible = true;
             cancelBtn.Visible = true;
+            unloadForm();
+           
         }
         private void loadForm()
         {
-            
+
             adddBtn.Visible = true;
             examBtn.Visible = true;
             updateBtn.Visible = true; ;
-            saveBtn.Visible = false ;
+            saveBtn.Visible = false;
             cancelBtn.Visible = false;
+            nameTxtBox.Enabled = false;
+            CMNDtextBOX1.Enabled = false;
+            femaleBtn.Enabled = false;
+            phoneTxtBox.Enabled = false;
+            noteCbbox.Enabled = false;
+            addressTextbox.Enabled = false;
+            birthDateTimePicker.Enabled = false;
+            danhSachBenhNhanDgv.AutoResizeColumns();
+
+        }
+
+        void unloadForm()
+        {
+            nameTxtBox.Enabled = true;
+            CMNDtextBOX1.Enabled = true;
+            femaleBtn.Enabled = true;
+            phoneTxtBox.Enabled = true;
+            noteCbbox.Enabled = true;
+            addressTextbox.Enabled = true;
+            birthDateTimePicker.Enabled = true;
+
         }
         private void saveBtn_Click(object sender, EventArgs e)
         {
-            if(isUsingAdd)
+            if (isUsingAdd)
             {
                 try
                 {
                     setValuefromForm();
-                    benhNhan_BUS.addBenhNhan(
-                                             fromForm.CMND,
-                                             fromForm.NGAYSINH,
-                                             fromForm.HOTEN,
-                                             fromForm.GIOITINH,
-                                             fromForm.DIACHI,
-                                             fromForm.SDT,
-                                             fromForm.GHICHU);
-                    showData();
-                    return;
-                }
-                catch(Exception)
-                {
-                    MessageBox.Show("Xảy ra lỗi trong quá trình thêm dữ liệu");
-                }
-            }
-            else if(isUsingUpdate)
-            {
-                try
-                {
-                    setValuefromForm();
-                    
-                    benhNhan_BUS.updateBenhNhan(fromForm.MABENHNHAN,
-                                             fromForm.CMND,
-                                             fromForm.NGAYSINH,
-                                             fromForm.HOTEN,
-                                             fromForm.GIOITINH,
-                                             fromForm.DIACHI,
-                                             fromForm.SDT,
-                                             fromForm.GHICHU);
-                    showData();
-                    return;
+                    if (checkData())
+                    {
+                        benhNhan_BUS.addBenhNhan(
+                                                 fromForm.CMND,
+                                                 fromForm.NGAYSINH,
+                                                 fromForm.HOTEN,
+                                                 fromForm.GIOITINH,
+                                                 fromForm.DIACHI,
+                                                 fromForm.SDT,
+                                                 fromForm.GHICHU);
+                        showData();
+                        loadForm();
+                        return;
+                    }
                 }
                 catch (Exception)
                 {
                     MessageBox.Show("Xảy ra lỗi trong quá trình thêm dữ liệu");
-
+                }
+            }
+            else if (isUsingUpdate)
+            {
+                try
+                {
+                    setValuefromForm();
+                    if (checkData())
+                    {
+                        benhNhan_BUS.updateBenhNhan(fromForm.MABENHNHAN,
+                                                 fromForm.CMND,
+                                                 fromForm.NGAYSINH,
+                                                 fromForm.HOTEN,
+                                                 fromForm.GIOITINH,
+                                                 fromForm.DIACHI,
+                                                 fromForm.SDT,
+                                                 fromForm.GHICHU);
+                        showData();
+                        loadForm();
+                        return;
+                    }
+                }
+                catch (Exception)
+                {
+                    MessageBox.Show("Xảy ra lỗi trong quá trình cập nhật dữ liệu");
                 }
 
             }
+            
         }
 
         private void searchBtn_Click(object sender, EventArgs e)
@@ -173,6 +220,97 @@ namespace GUI
             string temp;
             temp = CMNDtextBox2.Text;
             showSearchResult(temp);
+        }
+        private bool checkData()
+        {
+            if (string.IsNullOrEmpty(nameTxtBox.Text))
+            {
+                MessageBox.Show("Bạn chưa nhập họ tên bệnh nhân", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                nameTxtBox.Focus();
+                return false;
+            }
+            if (string.IsNullOrEmpty(CMNDtextBOX1.Text))
+            {
+                MessageBox.Show("Bạn chưa nhập số CMND của bệnh nhân", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                CMNDtextBOX1.Focus();
+                return false;
+            }
+            if (string.IsNullOrEmpty(phoneTxtBox.Text))
+            {
+                MessageBox.Show("Bạn chưa nhập số điện thoại bệnh nhân", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                phoneTxtBox.Focus();
+                return false;
+            }
+            if (string.IsNullOrEmpty(addressTextbox.Text))
+            {
+                MessageBox.Show("Bạn chưa nhập danh mục cho sản phẩm", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                addressTextbox.Focus();
+                return false;
+            }
+            if (IsNumber(CMNDtextBOX1.Text) == false)
+            {
+                MessageBox.Show("CMND không hợp lệ, CMND chỉ chứa chữ số ", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                CMNDtextBOX1.Focus();
+                return false;
+            }
+            if (IsNumber(phoneTxtBox.Text) == false)
+            {
+                MessageBox.Show("Số điện thoại không hợp lệ, SDT chỉ chứa chữ số", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                phoneTxtBox.Focus();
+                return false;
+            }
+            if (phoneTxtBox.TextLength < 9)
+            {
+                MessageBox.Show("Số điện thoại không hợp lệ, vui lòng nhập ít nhất 9 chữ số", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                phoneTxtBox.Focus();
+                return false;
+            }
+           
+            return true;
+        }
+        //Hàm kiểm tra có phải là số hay không 
+        private bool IsNumber(string exText)
+        {
+            Regex regex = new Regex(@"^[-+]?[1-9]*\.?[0-9]+[0-9]$");
+            return regex.IsMatch(exText);
+        }
+
+        private void nameTxtBox_Leave(object sender, EventArgs e)
+        {
+            TextInfo text = new CultureInfo("en-US", false).TextInfo;
+            nameTxtBox.Text = text.ToTitleCase(nameTxtBox.Text);
+        }
+
+        private void clearForm()
+        {
+            nameTxtBox.Text = null;
+            CMNDtextBOX1.Text = null;
+            femaleBtn.Checked = false;
+            phoneTxtBox.Text = null;
+            addressTextbox.Text = null;
+            noteCbbox.Text = null;
+        }
+
+        private void cancelBtn_Click(object sender, EventArgs e)
+        {
+            if (MessageBox.Show("Bạn có muốn hủy bỏ thao tác đang thực hiện", 
+                                "Thông báo", 
+                                MessageBoxButtons.OKCancel, 
+                                MessageBoxIcon.Warning) == System.Windows.Forms.DialogResult.OK)
+            {
+                loadForm();
+            }
+        }
+
+        private void addressTextbox_Leave(object sender, EventArgs e)
+        {
+            TextInfo text = new CultureInfo("en-US", false).TextInfo;
+            addressTextbox.Text = text.ToTitleCase(addressTextbox.Text);
+        }
+
+        private void danhSachBenhNhanDgv_RowPostPaint(object sender, DataGridViewRowPostPaintEventArgs e)
+        {
+            danhSachBenhNhanDgv.Rows[e.RowIndex].Cells["STT"].Value = (e.RowIndex + 1).ToString();
         }
     }
 }
